Fixed bugs in lindprompt.
Lindprompt's run command was unable to detect arguments given as input from the console. Fixed it by changing line 48 and 56.

Lindprompt's bath-fallback was flawed since it did not run commands with pipes properly. Fixed it by changing $(${myarr[@]}) to $(eval ${myarr[@]}). - must be checked if eval leads to a security problem