Update forkandopen.py
assert(int(fdlist[-1]) != -1) checks if the child PID's return -1. This would be an error with parent processes, but not with child processes.
"On success, the PID of the child process is returned in the parent, and 0 is returned in the child.  On failure, -1 is returned in the parent, no child process is created, and errno is set to indicate the error."